+++ ohos_nweb/include/nweb.h
@@ -338,7 +338,7 @@ class OHOS_NWEB_EXPORT NWeb : public std::enable_shared_from_this<NWeb> {
      *
      * @param accessibilityIdGenerator Accessibility id generator.
      */
-    virtual void PutAccessibilityIdGenerator(std::function<int32_t()> accessibilityIdGenerator) = 0;
+    virtual void PutAccessibilityIdGenerator(std::function<int64_t()> accessibilityIdGenerator) = 0;
 
     /**
      * Set the NWebHandler that will receive various notifications and
@@ -863,7 +863,7 @@ class OHOS_NWEB_EXPORT NWeb : public std::enable_shared_from_this<NWeb> {
      * @param accessibilityId The id of the accessibility node.
      * @param action The action to be performed on the accessibility node.
      */
-    virtual void ExecuteAction(int32_t accessibilityId, uint32_t action) const = 0;
+    virtual void ExecuteAction(int64_t accessibilityId, uint32_t action) const = 0;
 
     /**
      * Get the information of the focused accessibility node on the given accessibility node in the browser.
@@ -874,7 +874,7 @@ class OHOS_NWEB_EXPORT NWeb : public std::enable_shared_from_this<NWeb> {
      * @return true if get accessibility node info successfully, otherwise false.
      */
     virtual bool GetFocusedAccessibilityNodeInfo(
-        int32_t accessibilityId, bool isAccessibilityFocus, OHOS::NWeb::NWebAccessibilityNodeInfo& nodeInfo) const = 0;
+        int64_t accessibilityId, bool isAccessibilityFocus, OHOS::NWeb::NWebAccessibilityNodeInfo& nodeInfo) const = 0;
 
     /**
      * Get the information of the accessibility node by its accessibility id in the browser.
@@ -883,7 +883,7 @@ class OHOS_NWEB_EXPORT NWeb : public std::enable_shared_from_this<NWeb> {
      * @return true if get accessibility node info successfully, otherwise false.
      */
     virtual bool GetAccessibilityNodeInfoById(
-        int32_t accessibilityId, OHOS::NWeb::NWebAccessibilityNodeInfo& nodeInfo) const = 0;
+        int64_t accessibilityId, OHOS::NWeb::NWebAccessibilityNodeInfo& nodeInfo) const = 0;
 
     /**
      * Get the information of the accessibility node by focus move in the browser.
@@ -893,7 +893,7 @@ class OHOS_NWEB_EXPORT NWeb : public std::enable_shared_from_this<NWeb> {
      * @return true if get accessibility node info successfully, otherwise false.
      */
     virtual bool GetAccessibilityNodeInfoByFocusMove(
-        int32_t accessibilityId, int32_t direction, OHOS::NWeb::NWebAccessibilityNodeInfo& nodeInfo) const = 0;
+        int64_t accessibilityId, int32_t direction, OHOS::NWeb::NWebAccessibilityNodeInfo& nodeInfo) const = 0;
 
     /**
      * Set the accessibility state in the browser.
